refactor: moves watcher to deicated file in channel folder and wraps util in useCallback hook
Relates #422
